fixed non-updating issue
