+++ Quarto/core/src/com/mechwv/quarto/screens/EndingScreen.java
@@ -2,11 +2,17 @@ package com.mechwv.quarto.screens;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
+import com.badlogic.gdx.audio.Music;
 import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.graphics.Texture;
+import com.badlogic.gdx.graphics.g2d.TextureRegion;
+import com.badlogic.gdx.scenes.scene2d.Actor;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.ui.ImageButton;
+import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
+import com.badlogic.gdx.scenes.scene2d.utils.Drawable;
+import com.badlogic.gdx.scenes.scene2d.utils.TextureRegionDrawable;
 import com.badlogic.gdx.utils.viewport.StretchViewport;
 import com.badlogic.gdx.utils.viewport.Viewport;
 import com.mechwv.quarto.GameRoot;
@@ -21,6 +27,7 @@ public class EndingScreen implements Screen {
     private OrthographicCamera camera;
     private Viewport viewport;
     private int winner;
+    private Music SFXclick;
 
     public EndingScreen(final GameRoot game, int winner){
         this.winner = winner;
@@ -32,7 +39,21 @@ public class EndingScreen implements Screen {
         stage = new Stage(viewport,game.spriteBatch);
         Gdx.input.setInputProcessor(stage);
         this.game = game;
-        retry = game.gm.getRetry();
+
+        SFXclick = game.gm.getSFXClick();
+        Texture texture = game.assets.manager.get(game.assets.play_again);
+        TextureRegion textureRegion = new TextureRegion(texture);
+        Drawable drawable = new TextureRegionDrawable(textureRegion);
+        retry = new ImageButton(drawable);
+        retry.setPosition(( game.virtual_screen_width/2- retry.getWidth()/2),( game.virtual_screen_width/2+200));
+        retry.addListener(new ChangeListener() {
+            @Override
+            public void changed(ChangeEvent event, Actor actor) {
+                SFXclick.play();
+                game.gm.update();
+                game.setScreen(new MainMenuScreen(game,false));
+            }
+        });
         board = game.gm.getEnd_board();
         if (winner == 1)
             winning = game.gm.getPlayer_1_won();
@@ -40,7 +61,6 @@ public class EndingScreen implements Screen {
             winning = game.gm.getPlayer_2_won();
         if (winner == 3)
             winning = game.gm.getDraw();
-
             stage.addActor(retry);
     }
